Homework 06 - M4

+++ ecommerce-caratungu/src/users/users.repository.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './users.entity';
 import { Repository } from 'typeorm';
-import IUserDto from './usersDto';
-import { users } from '../dB/usersDB'
+import { CreateUserDto } from './dtos/CreateUser.dto';
+import { users } from '../dB/usersDB';
+import { LoginUserDto } from 'src/auth/dtos/LoginUser.dto';
 
 @Injectable()
 export class UsersRepository {
@@ -16,32 +17,45 @@ export class UsersRepository {
     const start = (page - 1) * limit;
     const end = start + limit;
     if (start >= users.length) {
-      return 'No se encontraron usuarios';
+      throw new HttpException(
+        `No existen usuarios para mostrar en la página ${page}`,
+        HttpStatus.BAD_REQUEST,
+      );
     }
     return users.slice(start, end);
   }
 
   async getUserById(id: string) {
-    const user = this.usersRepository.findOne({
+    const user = await this.usersRepository.findOne({
       where: {
         id,
       },
       relations: {
         orders: true,
-      }
+      },
     });
-    if (user) {
-      return user;
+    if (!user) {
+      throw new HttpException(
+        'No existe usuario con el ID especificado',
+        HttpStatus.BAD_REQUEST,
+      );
     } else {
-      // return 'No existe usuario con ese id';
+      return user;
     }
   }
 
-  async createUser(user: IUserDto) {
-    return await this.usersRepository.save(user);
+  async createUser(user: CreateUserDto) {
+    try {
+      return await this.usersRepository.save(user);
+    } catch (error) {
+      throw new HttpException(
+        error.message,
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
   }
 
-  async updateUser(id: string, uUser: IUserDto) {
+  async updateUser(id: string, uUser: CreateUserDto) {
     const user = await this.usersRepository.findOne({
       where: { id },
     });
@@ -53,12 +67,25 @@ export class UsersRepository {
       user.phone = uUser.phone;
       user.country = uUser.country;
       user.city = uUser.city;
+      await this.usersRepository.save(user);
+      return { message: 'Usuario actualizado', id };
     }
-    await this.usersRepository.save(user);
-    return { message: 'Usuario actualizado', id };
+    throw new HttpException(
+      'No existe usuario con el ID especificado',
+      HttpStatus.BAD_REQUEST,
+    );
   }
 
   async deleteUser(id: string) {
+    const userToDelete = await this.usersRepository.findOne({
+      where: { id },
+    });
+    if (!userToDelete) {
+      throw new HttpException(
+        'No existe usuario con el ID especificado',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     await this.usersRepository.delete(
       await this.usersRepository.findOne({
         where: { id },
@@ -67,28 +94,36 @@ export class UsersRepository {
     return `Usuario con id: ${id} eliminado`;
   }
 
-  async loginUser(userInfo) {
+  async loginUser(userInfo: LoginUserDto) {
+    
     const userLogin = await this.usersRepository.findOne({
       where: {
         email: userInfo.email,
+      },
+      select: {
+        name: true,
+        password: true,
       }
-    })
+    });
     if (userLogin) {
       if (userLogin.password === userInfo.password) {
         return `Login exitoso para el usuario ${userLogin.name}`;
       }
     }
-    return 'Email o password incorrectos';
+    throw new HttpException(
+      'Email o password incorrectos',
+      HttpStatus.BAD_REQUEST,
+    );
   }
 
-  async preloadUsers () {
-    const usersInDB = await this.usersRepository.find();
+  async preloadUsers() {
+    const usersInDB: CreateUserDto[] = await this.usersRepository.find();
     if (usersInDB.length === 0) {
       for (const user of users) {
         await this.usersRepository.save(user);
       }
       return 'Precarga de usuarios realizada';
     }
-    return 'Ya existen usuarios en la BD'
+    return 'Ya existen usuarios en la BD';
   }
 }
